Make 'core' the defaultPublishConfig + Remove unnecessary maven-publish code
